rest api authentication mechanizm is implemented

+++ Backend/controllers/feed.js
@@ -1,15 +1,29 @@
 const path = require('path')
 const fs = require('fs')
 const Post = require('../models/post')
+const User = require('../models/user')
 const {validationResult} = require('express-validator/check')
 
 
 exports.getPosts = (req, res, next) => {
-  Post.find().then(posts =>{
-    res.status(200).json({
-      message:'Fetched posts successfully.',posts:posts
+  const currentPage = req.query.page || 1;
+  const perPage = 2;
+  let totalItems;
+  Post.find()
+    .countDocuments()
+    .then(count =>{
+      totalItems = count;
+      return Post.find()
+        .skip((currentPage - 1) * perPage)
+        .limit(perPage)
     })
-  })
+    .then(posts =>{
+      res.status(200).json({
+        message:'Fetched posts successfully.',
+        posts:posts, 
+        totalItems:totalItems
+      })
+    })  
   .catch(err => {
     if(!err.statusCode){
       err.statusCode = 500;
@@ -55,23 +69,29 @@ exports.createPost = (req, res, next) => {
   const title = req.body.title;
   const content = req.body.content;
   const imageUrl = req.file.path.replace("\\" ,"/");
-  
+  let creator;
   const post = new Post({
     title:title,
     imageUrl: imageUrl,
     content:content,
-    creator:{
-      name:'Furkan'
-    },
+    creator:req.userId,
   })
   post.save()
     .then(result => {
+      return User.findById(req.userId)
+    })
+    .then(user =>{
+      creator = user;
+      user.posts.push(post);
+      return user.save()
+    })
+    .then(result =>{
       console.log('POST CREATED');
       res.status(201).json({
         message: 'Post created successfully!',
-        post: result
+        post: post,
+        creator:{_id: creator._id, name:creator.name}
       });
-
     })
     .catch(err => {
       if(!err.statusCode){
@@ -110,6 +130,12 @@ exports.updatePost = (req,res,next) =>{
         throw error;
       }
 
+      if(req.userId !== post.creator.toString()){
+        const error = new Error('not Authorized')
+        error.statusCode=403;
+        throw error;
+      }
+
       if(imageUrl !== post.imageUrl){
         clearImage(post.imageUrl)
       }
@@ -139,11 +165,23 @@ exports.deletePost = (req,res,next) => {
         error.statusCode=404;
         throw error;
       }
+
+      if(req.userId !== post.creator.toString()){
+        const error = new Error('not Authorized')
+        error.statusCode=403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId)
     })
     .then(result =>{
-      console.log(result);
+      return User.findById(req.userId)
+    })
+    .then(user =>{
+      user.posts.pull(postId);
+      return user.save()
+    })
+    .then(result =>{
       res.status(200).json({ message: 'Deleted post'})
     })
     .catch(err => {
